+++ src/modal-component/details-modal.js
@@ -19,22 +19,40 @@ class ModalDetails extends HTMLElement {
 
                 .details-modal {
                     background-color: rgb(255, 255, 255);
-                    width: 50%;
+                    width: 100%;
                     height: 65vh;
                     position: absolute;
-                    display: flex;
-                    z-index: 997;
+                    display: grid;
+                    grid-template-columns: 1fr 1fr;
                     left: 50%;
                     transition: 0.5s;
-                    border: 1px solid red;
                 }                
-   
                 
+                .slot-1 {
+                    order: 1;
+                }
+
+                .slot-2 {
+                    order: 2;
+                }
+
+                .details-modal.active .slot-1 {
+                    order: 2;
+                }
+
+                .details-modal.active .slot-2 {
+                    order: 1;
+                }
+
             </style>
 
             <div class="details-modal" id="detailsModal">
-                <slot name="information-details-modal"></slot>
+                <div class="slot-1">
+                    <slot name="information-details-modal"></slot>
+                </div>
+                <div class="slot-2">
                 <slot name="form-details-modal"></slot>
+                </div>
             </div>
         `;
