+++ src/modal-component/title-information.js
@@ -21,7 +21,7 @@ class TitleDetails extends HTMLElement {
                 display: flex;
                 flex-direction: column;
                 align-items: flex-start;
-                margin-bottom: 2rem;        
+                margin-bottom: 2rem; 
             }
